+++ test/integration/postgres/window_pagination_test.exs
@@ -72,7 +72,7 @@ defmodule Lotus.Integration.Postgres.WindowPaginationTest do
         ORDER BY name
         """,
         variables: [],
-        data_repo: nil
+        data_source: nil
       }
 
       assert {:ok, result} =
@@ -92,7 +92,7 @@ defmodule Lotus.Integration.Postgres.WindowPaginationTest do
         ORDER BY name
         """,
         variables: [],
-        data_repo: nil
+        data_source: nil
       }
 
       assert {:ok, result} =