+++ test/lotus/preflight_postgres_test.exs
@@ -155,7 +155,7 @@ defmodule Lotus.PreflightPostgresTest do
         ]
       ]
 
-      Lotus.Config |> stub(:rules_for_repo_name, fn _repo_name -> config end)
+      Lotus.Config |> stub(:rules_for_source_name, fn _repo_name -> config end)
       :ok
     end
 
@@ -202,7 +202,7 @@ defmodule Lotus.PreflightPostgresTest do
         deny: []
       ]
 
-      Lotus.Config |> stub(:rules_for_repo_name, fn _repo_name -> config end)
+      Lotus.Config |> stub(:rules_for_source_name, fn _repo_name -> config end)
       :ok
     end
 
@@ -234,7 +234,7 @@ defmodule Lotus.PreflightPostgresTest do
         ]
       ]
 
-      Lotus.Config |> stub(:rules_for_repo_name, fn _repo_name -> config end)
+      Lotus.Config |> stub(:rules_for_source_name, fn _repo_name -> config end)
       :ok
     end